+++ src/v2019C1/extensions/Function/extract-to.ts
@@ -1,9 +1,9 @@
 import { getOrCreateSubstationSectionRequiredStructure } from '@/v2019C1/helpers'
 
 import type { Scl } from '@/v2019C1/config'
-import type { Chain } from '@dialecte/core'
+import type { Chain, ExtensionsMethodParams } from '@dialecte/core'
 
-export function extractTo(params: Scl.MethodsParams<'Function'>) {
+export function extractTo(params: ExtensionsMethodParams<Scl.Config, 'Function'>) {
 	const { chain, contextPromise } = params
 
 	return async function (params: {
@@ -22,7 +22,22 @@ export function extractTo(params: Scl.MethodsParams<'Function'>) {
 			contextPromise: Promise.resolve(context),
 		})
 
-		const excludeFromFSDExtraction = [{ tagName: 'LNode' as const, scope: 'children' as const }]
+		const excludeFromFSDExtraction = [
+			// LNode children
+			{ tagName: 'LNodeInputs' as const, scope: 'self' as const },
+			{ tagName: 'LNodeOutputs' as const, scope: 'self' as const },
+			{ tagName: 'DOS' as const, scope: 'self' as const },
+			// Function children
+			{ tagName: 'FunctionSclRef' as const, scope: 'self' as const },
+			{ tagName: 'Variable' as const, scope: 'self' as const },
+			{ tagName: 'GeneralEquipment' as const, scope: 'self' as const },
+			{ tagName: 'ConductingEquipment' as const, scope: 'self' as const },
+			{ tagName: 'ProcessResources' as const, scope: 'self' as const },
+			{ tagName: 'PowerSystemRelations' as const, scope: 'self' as const },
+			// Common
+			{ tagName: 'Labels' as const, scope: 'self' as const },
+			{ tagName: 'BehaviorDescription' as const, scope: 'self' as const },
+		]
 
 		const functionTreeToClone = await sourceChain.getTree({
 			exclude: extension === 'FSD' ? excludeFromFSDExtraction : undefined,