+++ CHANGELOG.md
@@ -5,6 +5,34 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.11] - 2026-01-29
+
+### Fixed
+
+- fix handling of `Private` elements in the `afterClone`hook
+
+## [0.0.10] - 2026-01-29
+
+### Changed
+
+- update core to 0.0.9
+
+## [0.0.9] - 2026-01-29
+
+### Changed
+
+- `afterClone` hook: exclude empty `Private` element from cloning process
+
+## [0.0.7] - 2026-01-29
+
+### Added
+
+- `extractTo` extensions: extract any `SubFunction` to a new database at the right level (ie : `Substation`, `VoltageLevel` or `Bay`)
+
+### Changed
+
+- adjust the getTree filtering on `Function` method extractTo for `FSD`
+
 ## [0.0.6] - 2026-01-29
 
 ### Added
@@ -21,7 +49,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- `extractTo` extensions: extract any function to a new database at the right level (ie : `Substation`, `VoltageLevel` or `Bay`)
+- `extractTo` extensions: extract any `Function` to a new database at the right level (ie : `Substation`, `VoltageLevel` or `Bay`)
 
 ## [0.0.3] - 2026-01-29